cambie todo

+++ app/src/main/java/com/example/tiendahotwheels/model/Producto.kt
@@ -4,6 +4,7 @@ data class Producto(
     val id: String,
     val nombre: String,
     val descripcion: String,
-    val precio: Int,
+    val precio: Double,
     val imagen: String
 )
+
